Add files via upload
CpCachedFileManager is used to ensure that we do not go all the way to the MSSQL for every lookup. This saves a lot of network activity for all those tables that are used for show.
Any record looked up using the primary key will be cached for 60 seconds, or until an Update is encountered. 
Override the Individual File overrides in global.